refactor(web-file-preview): openLightbox 泛化 openViewer 保图片零回归（#293）
