+++ packages/dsh-web-file-preview/src/client/style.css
@@ -30,8 +30,11 @@
 .fwp-lbox-toolbar button{background:rgba(255,255,255,.16);color:#fff;border:1px solid rgba(255,255,255,.32);border-radius:6px;padding:4px 12px;font-size:13px;cursor:pointer;line-height:1.5}
 .fwp-lbox-toolbar button:hover{background:rgba(255,255,255,.3)}
 .fwp-lbox-stage{position:absolute;inset:0;display:flex;align-items:center;justify-content:center;overflow:hidden;cursor:zoom-out}
-.fwp-lbox-img{max-width:none;max-height:none;width:auto;height:auto;transform-origin:center center;transition:transform .1s ease-out;will-change:transform;cursor:grab;user-select:none;-webkit-user-drag:none;touch-action:none}
-.fwp-lbox.dragging .fwp-lbox-img{cursor:grabbing;transition:none}
+.fwp-lbox-content{max-width:none;max-height:none;width:auto;height:auto;transform-origin:center center;transition:transform .1s ease-out;will-change:transform;cursor:grab;user-select:none;-webkit-user-drag:none;touch-action:none}
+/* issue #293 E2：mermaid 克隆 SVG 初始尺寸按视口宽（≤90vw 且 ≤82vh），viewBox 保持纵横比，
+   初始渲染不溢出查看器 stage；图片路径（.fwp-lbox-content 无此约束）行为逐字节不变。 */
+.fwp-lbox-content-svg{max-width:90vw;max-height:82vh;width:auto;height:auto}
+.fwp-lbox.dragging .fwp-lbox-content{cursor:grabbing;transition:none}
 .fwp-rendered{overflow:auto;min-width:100%;box-sizing:border-box;width:100%}
 .fwp-rendered-md{font-size:14px;line-height:1.65;color:var(--dsw-alias-label-primary,#1c1e26)}
 .fwp-rendered-md h1,.fwp-rendered-md h2,.fwp-rendered-md h3,.fwp-rendered-md h4{margin:.6em 0 .3em;line-height:1.3}
